Pass the misspelling hash identifier to the spelling menu observer
This CL passes the hash identifier of the misspelling under the cursor to the
spelling menu observer. The spelling menu observer will use this identifier to
collect feedback about user actions on misspellings. For example, the user could
ignore spelling suggestions, add a word to their custom dictionary, or select
one of the spellcheck suggestions. This feedback will be used to improve
spellcheck suggestions.

BUG=170514

Review URL: https://chromiumcodereview.appspot.com/15616002