[telemetry] GTest output for PageTests.
If you're running a PageTest, --output-format can be none or gtest. If you're running a PageMeasurement, it can be buildbot, csv, or block. ("none" logs failures to stderr)

Split PageMeasurementResults; now the base class only deals with recording the results, and its derived class BuildbotPMR outputs in Buildbot format. CsvPMR derives from BuildbotPMR for now, to maintain existing behavior (where the default is "csv", but it also prints a summary in buildbot format, so it works for both humans and bots).

BUG=239323
TEST=./run_tests; ./run_gpu_tests --browser=system pixel_test page_sets/pixel_test.json; ./run_multipage_benchmarks --browser=system smoothness_measurement page_sets/top_25.json
R=tonyg@chromium.org

Review URL: https://codereview.chromium.org/15522002