Added a PolicySchema class.
A PolicySchema describes the known policy keys (for Chrome, or for an extension)
and maps them to their expected base::Value::Type. This class parses a buffer
as a JSON schema and additionally verifies that it can be used to describe a
policy map (e.g. the schema version is defined, the top-level is an "object",
and other checks).

BUG=108992

Review URL: https://chromiumcodereview.appspot.com/15133005