Files.app: Fixed a bug of moving multiple files.
In the situation that both a directory and its child item are moved, if the
directory is moved first, we will lose the child item because its path is
changed.

This CL make the copy items sorted so as that items in the deeper directories
moved faster than ones in the shallower directories.

BUG=230755
TEST=Enable moving from search results and test manually

Review URL: https://chromiumcodereview.appspot.com/15806002