Handle errors when Files.app goes offline while there are images in the image loader's queue.
After Files.app goes offline, sending a response with a loaded image caused throwing an exception, which caused not releasing a task from the queue. As a result such tasks got stuck in the queue, which has limited capacity. This situation caused clogging the queue, which made it impossible to handle further tasks. This patch resolves this issue by ignoring an error in sending the response.

TEST=Tested manually. After entering and leaving the gallery several times, images still work.
BUG=239237
R=yoshiki@chromium.org

Review URL: https://codereview.chromium.org/15845002