Added a PolicyDomainDescriptor.
Policies are scoped by namespace keys, which indicate whether they're meant for
the browser itself or for a component, such as an extension. The descriptor
additionally describes the types of the known policies, for each component. This
allows the policy loaders to known which policies to look for, and to filter out
unknown or invalid policies.

BUG=108992

Review URL: https://chromiumcodereview.appspot.com/15061007