Remove trailing commas from Chromium IDLs
This removes trailing commas in extended attributes from the
2 files that have them.
This is not part of the standard, and doesn't work with strict
parsers, but laxer parsers may not notice.
This removes a barrier to using more standard-compliant parsers
more widely; discovered when rewriting Blink IDL parser.

BUG=242800

Review URL: https://chromiumcodereview.appspot.com/15679004