First stages of the new add-limited-user flow: signing in and "Learn more"
Modify the create-profile dialog to only allow creation of a managed (aka limited) user
if the current user is signed in, and give an informative message otherwise.

Add a "Learn more" link to the checkbox for creating a limited user, and a descriptive
overlaythat appears when it is clicked.

BUG=234991
TEST=manual

Review URL: https://chromiumcodereview.appspot.com/14743018