Remove races in window creation in Files.app.
This patch removes races in case of requested reusing of windows. It might happend, that windows are not yet fully initialized and we tried to reuse them. This bug is fixed by waiting for all windows to be initialized by enqueuing the reusing logic.

TEST=Follow the bug's repro steps.
BUG=238412

Review URL: https://chromiumcodereview.appspot.com/15817002