Postpone the list and the grid views' resizing.
When having lots of items on the file list, resizing of the window became very sluggish. This patch delays resizing of the file list and the grid view by 100ms, so the rest of the UI can be rendered very fast. The drawback is that the file list and grid view are updated with noticeable delay, but it looks much better than before.

TBR=yoshiki@chromium.org
TEST=Go to a directory with > 40 files. Resize the window.
BUG=242965

Review URL: https://chromiumcodereview.appspot.com/15716003