[content shell] Fix layout test system dependencies check on Windows
The problem is that "run_webkit_tests.py --nocheck-sys-deps" on Windows does
not work as excepcted. Because, content shell always executes the layout test
system dependencies check. On the other hand, DumpRenderTree executes the
dependencies check if and only if --check-layout-test-sys-deps option is specified.
As you can see, content shell and DRT behave differently on this point.
This CL fixes this problem.

BUG=111316

Review URL: https://chromiumcodereview.appspot.com/15732007