Move the OmniboxViewViews font-selection algorithm to LocationBarView and use the passed-in font in OmniboxViewViews.  This does not change the appearance of OmniboxViewWin in the default case; it might have small effects if the user has a non-default system font/size, but this code is about to be changed again for the large omnibox, so I don't care too much.
This does have a small effect on OmniboxViewViews when the user has a
larger-than-normal system font: instead of "exactly" 15/12 px, we'll now do "at
least" those sizes, and then clamp to what will fit in the omnibox.  This
matches the OmniboxViewWin behavior and seems generally more correct.

Also eliminate OmniboxView::GetFont(), which is unused as of my previous patch
to location_bar_view.cc.  I didn't mean to eliminate the use of this at the
time, and doing so introduced a potential disagreement in the
|suggested_text_view_| font and the OmniboxViewViews font; this patch resolves
that problem, and is where I'd intended to kill the API to begin with.

BUG=231005
TEST=Omnibox fonts should not look different in the common case
R=sky@chromium.org

Review URL: https://codereview.chromium.org/15848004