Safety check for single-token commands.
This is a fix for the problem seen here:
http://build.chromium.org/p/client.quickoffice/builders/Master%20Viewer/builds/242/steps/Install%20Build%20Deps/logs/stdio
A single string token was passed as the command, rather than a list/tuple of
strings, and so python did its iterable magic. We'd like to prevent that so
people don't shoot themselves in the foot.

R=iannucci@chromium.org, xusydoc@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/16033004