RulesRegistryWithCache::ProcessChangedRules speed-up
Postpone rules->value conversion:
Currently the rules are converted to base::Value on the rules registry thread before ProcessChangedRules finishes, and then a task is posted to UI which will trigger writing that to a rule store on yet another thread.
This CL postpones the conversion to the task running on UI thread.

This change cuts down the running time for AddRules to 75%, but of course this is cheating, since the work only gets done later. The point is, however, that network requests are not blocked until the rules are written.

BUG=236368

Review URL: https://chromiumcodereview.appspot.com/14851011