setCookie shouldn't convert WebString to 16 bits
Previously, setCookie forced WebString to have a 16 bit backing buffer by
calling WebString::data() in the common case, that means we'd create and retain
a 16 bit version of the string in addition to the 8 bit version. We're working
to remove WebString::data() from the API because it's a memory footgun.

In profiling Mobile Gmail, this line of code triggered 50% of the remaining 16
bit upconversions (after my other in-flight changes land).

R=jamesr
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/16022010