Add a lot of docstrings and remove 3 lines of code.
Remove the check for 'testfilter' presence in
FactoryCommands.TestStepFilterImpl(). I manually verified 'non_default' is not
improperly used on non-TS server. As of this comment, 'run_default_swarm_tests' is
set on the following builders:
- Try Server for linux_rel and win_rel only.

In particular, the fact it was set on chromium.swarm at all was an error, since
this is specifically for builder/tester kinds of setup, and chromium.swarm uses
a single builder setup instead. This was an error introduced in r202340.

TBR=iannucci@chromium.org
BUG=242716

Review URL: https://chromiumcodereview.appspot.com/16049003