Made the open menu item in the context menu updated when the context menu is open.
Originally, FileSelectionHandler#updateFileSelectionAsync have the complex
conditional statements. And when util.platform.newUI() is true and
selection.totalCount is zero, updateContextMenuActionItems that updates the open
menu item is not called.

This CL arranged the statements by the target of update and added some comments.

BUG=243687
TEST=manually

Review URL: https://chromiumcodereview.appspot.com/15950003