Add a policy list for access to capture devices.
The approach is slightly different from my proposal in the bug so that we have more configuration options as well as backwards compatibility with Chrome M25.

This change introduces two new policy lists that allow admins to whitelist URLs to get no-prompt access to capture devices via URLs or URL patterns.
The pattern matching is applied only to the security origin of the URLs.

The behavior of the existing AudioCaptureAllowed and VideoCaptureAllowed is reverted back to how it was prior to r180416.  This means that admins must use the whitelist to allow device access without prompt.

If no match for a URL is found in the whitelists, the default behavior is determined by the [Audio|Video]CaptureAllowed policy value.

BUG=225045

Review URL: https://chromiumcodereview.appspot.com/15738004