[content shell] Do not disable cc-driven animation when using threaded compositing
The kDisableThreadedAnimation flag, which actually disables cc-driven animations
regardless of whether we're using a compositor thread, is currently set for all
layout tests. We need to set this flag in single thread mode (which is the default
for layout tests) since, in this mode, cc-driven animations cause never-ending
invalidation, causing tests to fail. However, when threaded compositing is enabled,
cc-driven animations do not cause never-ending invalidation.

This CL makes us only disable cc-driven animations when in single thread mode.
This is a prerequisite for creating a virtual test suite for animations that uses
threaded compositing and threaded animation.

BUG=244266

Review URL: https://chromiumcodereview.appspot.com/15939004