Fix app gl state restore
GL_ARRAY_BUFFER and GL_ELEMENT_ARRAY_BUFFER could definitely be
non-zero, as seen in testing. So always restore 0. Also move them to the
last things to be restored. This patch seem to be helping a lot with not
destroying app UIs.

BUG=230195
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/16096004