linux_aura: Close window when the OS requests it.
Alt-F4 key events are handled by the window manager, who then sends an X11
message telling us that we should close. This used to be hooked up, but it
looks like it broke when we merged with the crossplatform
DesktopNativeWidgetAura code that we also use on Windows.

BUG=176783

Review URL: https://chromiumcodereview.appspot.com/15827009