Reduce usage of WebString::data (Chromium side)
WebString::data forces a widening of the underlying String from 8 bits to 16
bits. We should remove this function because most callers of the API don't
understand these implementation details of WebString. In order to remove the
API, we need to remove all the callers. This CL removes all the callers I could
find in Chromium. This CL depends on https://codereview.chromium.org/15806014/
which removes the callers in Blink (and introduces WebString::at to address one
data() use case).

R=jamesr
TBR=groby@chromium.org

Review URL: https://chromiumcodereview.appspot.com/15960012