Improve recommended sample counts for PPAPI.
Testing shows we can allow for much lower buffer sizes on WASAPI,
OSX, and PulseAudio clients.  In most cases, this change brings those
clients down to the nearest multiple of 512.  WaveOut and ALSA clients
will use the nearest multiple of 2048.

Based on testing done in https://codereview.chromium.org/14570002/

BUG=162207
TEST=Using these buffer sizes for HTML5 audio works without glitching
when an AudioConverter is in the graph.

Review URL: https://chromiumcodereview.appspot.com/14573006