Add enable-app-launcher-shim to chrome://flags.
Adds a flag for --enable-app-launcher-shim to chrome://flags. Enable to 
ensure an OSX app bundle for the App Launcher exists under the 
Applications folder. If the flag is not present any previously created 
app bundle for the App Launcher will be deleted on startup. 

BUG=138633
TEST=Tested enabling and disabling the flag in chrome://flags.

Review URL: https://chromiumcodereview.appspot.com/15675008