Add CompositingReasons to compositor layers.
This is a re-land after first version (r202580) was reverted. The problem was
that extra invalidations occurred due to a call to SetNeedsCommit(). The
solution is to avoid calling SetNeedsCommit(), and accept that this debug info
will be stale on the impl-thread until something else commits.

BUG=240946

Review URL: https://chromiumcodereview.appspot.com/16023024