Run presubmit without --verbose
This code originally passed --verbose only when CQ was 
started with --verbose, but after r181360 it was always
passed. This logic used the global logging level to decide
what arguments to pass to presubmit, and the dependency
was missed when CQ logging was expanded and the
meaning of the global logging level changed.

Review URL: https://chromiumcodereview.appspot.com/15812005