Enable strace using optionally sudo.
With sudo usage, it is easiest (guaranteed) that the process can be attached,
so use an exec thunk to start the child process. This gives a deterministic
process tree that is easier to parse. Keep the non-sudo functionality which is
fine enough for the vast majority of cases.

R=csharp@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/15951010