Correct immersive mode gestures under rotation
This code does to things to make sure that immersive mode gestures operate
correctly regardless of the screen orientation. The first is to move the
registration of the filter to the root window, which guarantees that all of the
events will be seen by the filter.

The second change is to extend the bounds the filter is watching above by the
maximum bezel side. This makes sure that under rotation swipes that start on the
bezel are still handled.

BUG=241545, 244269
TEST=Rotated screen through all rotations and confirmed that swipe in/out works
     as expected in immersive mode.

Review URL: https://chromiumcodereview.appspot.com/15734011