Moved DOMActionType information from extras into a real field (ActivityLog)
Previously, the DOMActionType was being sent from the renderer to the browser ActivityLog as a string. It was then being turned into an enum. I've now moved it to its own proper field, and it's being marshaled across the renderer/browser boundary as an int that matches an enum instead of as a string.

R=mpcomplete@chromium.org
BUG=229703

Review URL: https://chromiumcodereview.appspot.com/15520002