cc: Ignore offscreen tiles for activation
Remove the overenthusiastic DCHECK from PictureLayerImpl to make sure
that every tile that intersects the visible content rect also has a
distance to the viewport of zero.  Because the projection of the
viewport is done more precisely than the enclosing rect for the visible
content rect, it may be possible for a tile to be outside the viewport
but inside the visible content rect.

Therefore, the current DCHECKS are bogus.  Instead, skip these tiles
because they're not truly required for activation.

R=reveman@chromium.org
BUG=243437

Review URL: https://chromiumcodereview.appspot.com/16160005