Reduce wait time for Windows in AudioOutputController:WaitTillDataReady()
This timeout was long to work around a problem that has been resolved
elsewhere.
I'm also changing the histogram name to match the condensed format
that is now preferred. In addition I've added a new boolean histogram
to indicate if the data was ready or not.

BUG=244939
BUG=242726

TEST=None

Review URL: https://chromiumcodereview.appspot.com/16189009