net/disk_cache/simple: Reduce size of EntrySet nodes.
The EntrySet type was a base::hash_map that duplicated the hash_key
value in both the key and value types, wasting 8 bytes per node.

Changing the implementation to a base::hash_set has been experimented
with and introduces lots of non trivial code changes (some of them being
dependent on the exact C++ STL implementation).

This is a simpler optimization that simply puts the hash_key out of the
EntryMetadata structure, and updates all related code appropriately.

Note that this only changes the in-memory representation of the index,
the on-disk format is untouched.

BUG=

Review URL: https://chromiumcodereview.appspot.com/14746019