cros: Ignore ERR_ABORTED to break potential infinte loading loop.
ERR_ABORTED could be triggered from LoadAuthExtension() and ReloadGaiaScreen()
if there is pending load on gaia frame and new load request would cause that
pending load to be aborted. Infinite loading loop could be triggered when the
test URL that triggers the test content script to auto submit the login form is
loaded. The scenario is roughly like the following:
1. When the test URL is loaded with an error screen (frame error 11), the
   login form loaded okay and a POST request is created by content script.
2. The form loading okay clears the error and ReloadGaiaScreen() follows that.
3. When POST comes back from server, whola, ERR_ABORTED is generated because
   that request is aborted;
4. ERR_ABORTED brings up the error screen and a ReloadGaiaScreen() is called
   to retry the loading;
And we go back to step 1 and this could go on and on.

This CL breaks the loading loop by ignore ERR_ABORTED once
per LoadAuthExtension/ReloadGaiaScreen call.

BUG=242527
R=ygorshenin@chromium.org, zelidrag@chromium.org

Review URL: https://codereview.chromium.org/15876012