Determine card egilibility by card.status from server only.
Sugar server already set status when the card is not eligible for transaction, like UNSUPPORTED_COUNTRY, AMEX_NOT_SUPPORTED etc.

This CL also enables BILLING_INCOMPLETE as we want it to be selectable and put it under edit mode when the user selects it.

BUG=245006

Review URL: https://chromiumcodereview.appspot.com/16208002