Fix for closing idle SPDY sessions with aliases to unblock a lower level stalled socket pool.
Issue was that SpdySession::CloseOneIdleConnection() would
return false, despite deleting the session.  Then the caller
would try and follow a no-longer-valid iterator to the next
layered socket pool.  Fix is to have the function return
true in that case.

This was broken in https://codereview.chromium.org/199880

BUG=244730

Review URL: https://chromiumcodereview.appspot.com/16229002