Move C++ code from webkit/tools/test_shell to webkit/support
This code is simple/test/mock implementations of APIs to use in DumpRenderTree,
webkit_unit_tests, and various other bits of unit test code around the codebase.
Nothing here is used for test_shell since there is no test_shell any more.

This moves the code to webkit/support/ along with the rest of DRT/w_u_t support
code. Hopefully we'll be able to delete most of this soon along with
DumpRenderTree. In the meantime this simplifies webkit/

R=darin@chromium.org

Review URL: https://codereview.chromium.org/15946006