Handle DemuxerStream config change in VideoFrameStream.
This CL handles DemuxerStream::kConfigChanged in VideoFrameStream and hides it from the downstream VideoDecoders. So after this CL, kConfigChanged handling code in all VideoDecoder implementations can be deleted!

Once kConfigChanged is detected, the internal state switches to STATE_FLUSHING_DECODER. The VideoFrameStream flushes the VideoDecoder before trying to reinitialize it. In this state, the VideoFrameStream stops pulling any data from the real DemuxerStream and always feeds the VideoDecoder with EOS buffer. After the VideoDecoder is flushed, it is reinitialized with the new config.

It's a bit tricky to handle the Reset() call during this process. Once the VideoFrameStream is in the STATE_FLUSHING_DECODER state, it has to finish the VideoDecoder reinitialization process so that it is ready for the next ReadFrame() call.

BUG=141788
TEST=Current tests passes. Tested normal playback on Window/Linux. Tested a lot of config changes with MSE/EME.

Review URL: https://chromiumcodereview.appspot.com/14779008