Menu.js: Not to set a text of menu item when the command label is not specified.
The label of the menu item is automatically set by the corresponding command label. But, previously, when the command label was not specified and the corresponding menu item has a text label, the text label was overridden by an empty text. This patch prevents it. 

BUG=242270
TEST=manual

Review URL: https://chromiumcodereview.appspot.com/15397008