Move synchronous compositor into content/browser
This means all the hookup can remain in content/ so less internal
details need to be revealed in the public API long-term.
Now there is a single API to intialize the sync compositor mode per
process, and per WebContents, rather than a collection of flags and
hooks. The bulk of the browser->renderer hookup is done in content/ so
android_webview InProcessViewRenderer can be greatly simplified.

(depends on pending CL https://codereview.chromium.org/15875009/)

Review URL: https://chromiumcodereview.appspot.com/15851006