Shows Omnibox autocompletion when IME on.
**** NOTE: This CL was cloned from codereview.chromium.org/13952009

Shows the inline autocompletion in Omnibox even when IME is turned on. 
This CL adds a new textfield to show the inline autocompletion so it will not affect (or be affected by) a IME, leaving the Omnibox textfield under the control of IME. 

The way to hide the Omnibox popup window when IME popup is open is changed. The old way is just to stop requesting autocompletion. The new way is to close the Omnibox popup but continue requesting autocompletion. 

This CL supports only ChromeOS and Windows.  I'll support other platforms later. 

How to see the improvement with this CL: 
1. Visit gmail.com several times by typing it so Chrome suggests gmail.com when you type [gm]. 
2. Enable Japanese IME. 
3. Type [g] or [gm]. Preedit is [g] or [gm], or its full-width versions. 

Old behavior: No inline autocompletion is shown. 

New behavior: [g]mail.com or [gm]ail.com is suggested as the inline autocompletion, with the normal text color (black by default) and focused selection background color (blue by default). 

4. Hit enter key. Preedit is committed. 

Old and new behavior (no change here): [g]mail.com or [gm]ail.com is suggested as the inline autocompletion, with the normal text color (black by default) and focused selection background color (blue by default). 

5. Hit enter key. You visit gmail.com. 

BUG=159105

Review URL: https://chromiumcodereview.appspot.com/14219006