Files.app: Keep the total width of table when resizing a column by mouse drag.
This CL added a layout method setWidthAndKeepTotal to FileColumnModel and
extends TableSplitter class so that it calles the FileColumnModel's layout
method.

In this layout method, the width of specified column is changed and the width of
columns at the next of the resized column are adjusted so that it is keeping the
total width of table.

BUG=238867
TEST=manually

Review URL: https://chromiumcodereview.appspot.com/16114014