Fixing a race condition in SpeechRecognizerImpl.
In very rare cases, a SpeechRecognizerImpl can receive an abort
request followed by a start request after the session has been
created (by CreateSession). The abort request causes the
SpeechRecognitionManagerImpl to de-refcount, for that particular
session, the SpeechRecognizerImpl, which in turn will be freed
soon after serving the start request (posted tasks keep the
SpeechRecognizerImpl refcounted as long as they're in the message
loop).
The SpeechRecognizerImpl, however, is an event handler for the
AudioInputController, which unfortunately uses a naked pointer.
Furthermore, the AudioInputController handles most of its operations,
including Close, asynchronously on another thread.
This fix maxes unfeasible to start a recognition session after abort() is
called, ensuring that no further operations can be invoked on the
SpeechRecognizerImpl after the OnRecognitionEnd event.

BUG=244415
TEST=content_unittests --gtest_filter='*Speech*'

Review URL: https://chromiumcodereview.appspot.com/15812006