Identity API: switch WebAuthFlow dialog to component app
All Identity API UI is moving into a webview tag inside a component
app, allowing Chrome Apps to be isolated from the browser.

This change updates WebAuthFlow, which controls the dialogs, to invoke
the component app instead of a browser pop-up.

BUG=228908
(This is part of step #6 in the description.)

Review URL: https://chromiumcodereview.appspot.com/15897006