gpu: Autogenerate glHint state
This patch implements automated code generation for state that is
set with the glHint function. Currently this state includes
GL_GENERATE_MIPMAP_HINT and GL_FRAGMENT_SHADER_DERIVATIVE_HINT_OES.

In order to support functions like glHint that take an enum identifying
the state variable whose value is being set, we add a new
'NamedParameter' state type to the command buffer code generator.
Functions of this type are expected to take the parameter enum as the
first argument and the value as the second one. This same mechanism can
also be used with glPixelStorei and other similar functions.

BUG=245228
TEST=glHint test in [1] passes repeatedly with --enable-virtual-gl-contexts.

[1] https://www.khronos.org/registry/webgl/conformance-suites/1.0.1/conformance/extensions/oes-standard-derivatives.html

Review URL: https://chromiumcodereview.appspot.com/15792007