Identity API: switch WebAuthFlow dialog to component app
This is a reland of 203417, which should work now that 203506 is in.

All Identity API UI is moving into a webview tag inside a component
app, allowing Chrome Apps to be isolated from the browser.

This change updates WebAuthFlow, which controls the dialogs, to invoke
the component app instead of a browser pop-up.

TBR=miket@chromium.org, mpcomplete@chromium.org

BUG=228908
(This is part of step #6 in the description.)

Review URL: https://codereview.chromium.org/15874006