cc: Add new RasterWorkerPool interface.
Refactor RasterWorkerPool interface such that it can be used
to implement different type of resource updates mechanisms.
The result is that the tile manager is no longer aware of
the need to perform uploads. This cleans up the tile manager
code significantly and allow zero copy resource updates to
be properly supported.

BUG=241644
TEST=cc_unittest --gtest_filter=BasicRasterWorkerPoolTest.*

Review URL: https://chromiumcodereview.appspot.com/16190002