Differentiate between inputs, outputs and touched files on OSX and linux.
It is not functional yet on Windows, and it's not guaranteed it's possible at
all so mark all the files as modified for now.

This is to eventually be able to trace meaningfully ninja, at least on OSX and
linux.

Fix a regression to run_isolated_smoke_test.py on Windows due to r203426.

R=csharp@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/16240015