Create install/uninstall browser tests for High DPI Theme extensions
Adding a new directory that will store the *.crx file and the private key
file for (the recently added) test extension for High DPI themes.
For now I only add the private key file (text). The binary *.crx file
will be added once this has landed (via dcommit).

Note this new directory is in parallel with the existing dir for the
extension. This is because the private key file should not be stored
with the extension otherwise packing the extension will fail.

New dir:

  .../chrome/test/data/extensions/theme_hidpi_crx

Existing dir (with extension):

  .../chrome/test/data/extensions/theme_hidpi

BUG=245843
R=sky@chromium.org, asargent@chromium.org
TEST=manual (and future install/uinstall test)

Review URL: https://chromiumcodereview.appspot.com/16154011