Make resizing in Files.app more responsive.
Before, the resizing of the file list and the grid list was delayed to avoid making the UI janky. That delay caused flickering of the screen when showing the Save-as dialog as well as when maximizing/restoring the window.
This patch solves this problem by calling immediately resizing the file list and grid view, but delaying any consecutive resizes.

TEST=Maximize, minimize, check the Save-as dialog.
BUG=245042

Review URL: https://chromiumcodereview.appspot.com/15809008