Files.app: Added DragSelector class.
DragSelector class enables user to select multiple files by dragging.  By
calling DragSelector#startDragSelection(event), drag selection is
started. DragSelector class added handlers of mouse events, shows drag handler
and selects files by the mouse operation.

In order to call startDragSelection, users of DragSelector class must handle
'dragStart' event of the list and decide whether to call startDragSelection or
not according to the event. (e.g. whether the start point is hit to background
of the list or not)

If the drag selection is started with modifier keys (shift or ctrl),
DragSelector stores the selection state at the beginning of drag selection. If
an item included in drag handle once and then it is get out of the handler, the
selection state of the item is restored by the stored initial selection state.

BUG=224832
TEST=Added a call of startDragSelection to FileTransferManager, and
do drag operation on the file list.

Review URL: https://chromiumcodereview.appspot.com/15643006