File system API: whitelist the Downloads dir on Chrome OS.
Due to crrev.com/203393 we now blacklist the user data dir in
DoCheckWritableFile(). However, on Chrome OS the user Downloads
folder (~/Downloads) is a subdirectory of the user data dir and
serves the role of local storage. This should be writable for
file system API.

BUG=246339
TEST=verified Downloads folder is writable by v2 apps on Chrome OS.

Review URL: https://chromiumcodereview.appspot.com/15896035