Clear Reference Location State When Providers are Stopped
The location arbitrator relies on pointer comparisons to identify
which location provider provided a reported location.

Now, we clear this state when the providers are shut down.

This prevents a bug where a stale location prevented the arbitrator
from updating its clients with a newer location.

BUG=240956

Review URL: https://chromiumcodereview.appspot.com/15968009