Fixed extension/app window maximum size calculation on Windows.
Previously the height of window title bars was added to the maximum
content size of windows only if there was a maximum width as well as a
maximum height. This fixes that so it's possible to specify maxHeight
without specifying maxWidth to chrome.app.window.create().

BUG=246289

Review URL: https://chromiumcodereview.appspot.com/15969028