Make WebRTC Linux builders not share machines with CrOS builders.
It is more important for us to have higher frequency on the Linux
builders than CrOS. Because of this I'm moving the daisy CrOS builder to
the machine that already runs x86 and amd64 CrOS builds.

I have verified the machines that will take on additional checkouts have
enough disk space to drive this.

BUG=none
TEST=presubmit checks

Review URL: https://chromiumcodereview.appspot.com/16286007