Extensions: pass ChromeV8Context around instead of v8::Handle.
There was a TODO about this in object_backed_native_handler.h.

The passed handle was a Persistent behind the scenes, and we're going to disable
copying Persistents. We'd need to pass a Local handle instead, but this way the
TODO gets done with a similar amount of effort.

BUG=236290

Review URL: https://chromiumcodereview.appspot.com/16032015