RulesRegistryWithCache::ProcessChangedRules postponed
The main change here is that ProcessChangedRules (includes converting the rules
and writing them to disk) is now postponed to when the rules registry becomes
ready. After start-up, this will just mean that ProcessChangedRules will be
added to the message loop queue.

A minor change is a new way to manage ProcessChangedRules tasks:
If there is not the UI part (which does the writing), ProcessChangedRules tasks
are not ever posted.
If there is the UI part, then at any time at most one ProcessChangedRules task
is ever scheduled. This avoids unnecessary work duplication.

BUG=236368

Review URL: https://chromiumcodereview.appspot.com/15911005