Don't track whether a resize ack is pending during layout tests
Unlike normal operation, in layout tests the renderer can request a synchronous
change of its size. The renderer and the browser therefore often run out of
sync. However, it's important that the renderer always acks a resize request
from the browser.

Previously, I tried to track the size the browser expects the renderer to have
in the renderer, however, this turns out to not work when the compositor is in
hidpi mode. Instead of trying to fix this up somehow, just ignore whether or
not the browser is expecting resize acks during layout tests

BUG=244369

Review URL: https://chromiumcodereview.appspot.com/16285005