Make sure OneClickSigninSyncStarter detects Browser object being closed
Before this fix, OneClickSigninSyncStarter didn't detect when the Browser object it points to was closed. It could then cause a crash by making calls on the destroyed Browser object. OneClickSigninSyncStarter now observes the browser list to detect when the Browser it points to is closed.

BUG=235439

Review URL: https://chromiumcodereview.appspot.com/16092009