Disable dcheck on layout release bots.
Currently the layout_rel bots are configured to enable dcheck; this is
good in that it gives us DCHECK coverage without having to run the slow
debug bots, but bad in that if we have a flaky test that fails DCHECKS, it
will fail on the try servers but not on the main bots. Since we don't
capture flakiness results from the try servers, this means we end up
getting "unexpected" flakiness.

This may be contributing to the vastly increased flakiness in the CQ.

This patch turns this configuration off for now; we can turn it back on
if we get to a point where we can handle flakiness other ways.

R=iannucci@chromium.org, thakis@chromium.org, eseidel@chromium.org
BUG=238381

Review URL: https://chromiumcodereview.appspot.com/16097011