Empty rect should not intersect with another rect
Checks if the rect is empty in Rect::Intersects() in order to
1. keep consistence with IntRect in blink;
2. avoid (0,0 0x0) from intersecting rects containing the origin;
2. avoid unnormalized empty rects from intersecting other rects.

This will reduce some inner-loop computations e.g. in iterators of
cc:PictureLayerTiling.

However, gfx::RectBase::Contains(empty_rect) is still in question.
We should determine what the correct behavior is and fix it in the code.

BUG=246760

Review URL: https://chromiumcodereview.appspot.com/14109023