chromeos: Handle new PowerSupplyProperties fields.
This updates PowerManagerClient to handle changes to
PowerSupplyPropreties protocol buffers (namely, the addition
of an ExternalPower enum and the replacement of the existing
BatteryState enum). It just maps the updated fields to
Chrome's existing PowerSupplyStatus struct.

It also updates the system_api dependency to f893cdef.

BUG=chromium:245108

Review URL: https://chromiumcodereview.appspot.com/16338003